39 refactor backend users

+++ CodeKickFc/backend/src/main/java/com/codecool/codekickfc/controller/dto/users/UserMatchDTO.java
@@ -1,4 +1,4 @@
 package com.codecool.codekickfc.controller.dto.users;
 
-public record UserMatchDTO(int userId, int matchId) {
+public record UserMatchDTO(long userId, long matchId) {
 }
